Reorganized file paths and locations
