+++ src/main/java/com/projectshowdown/configs/RandomDataController.java
@@ -1,4 +1,4 @@
-package com.projectshowdown.firebase;
+package com.projectshowdown.configs;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
