+++ src/main/java/com/projectshowdown/configs/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.projectshowdown.security;
+package com.projectshowdown.configs;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
