+++ src/main/java/com/projectshowdown/controllers/PlayerController.java
@@ -1,4 +1,4 @@
-package com.projectshowdown.player;
+package com.projectshowdown.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -10,6 +10,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.projectshowdown.exceptions.PlayerNotFoundException;
+import com.projectshowdown.player.Player;
+import com.projectshowdown.service.PlayerService;
 
 import jakarta.validation.Valid;
