+++ src/main/java/com/projectshowdown/exceptions/PlayerNotFoundException.java
@@ -1,4 +1,4 @@
-package com.projectshowdown.player;
+package com.projectshowdown.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
